removed unused layer design panels and added icons to shared design type values
