+++ app/routes/sketch+/artboards+/$slug_+/components/panel/design/type/panel-design-type-row-values-rotate.tsx
@@ -3,6 +3,7 @@ import {
 	SidebarPanelRowValuesDisabled,
 } from '#app/components/templates'
 import { type IRotate } from '#app/models/rotate.server'
+import { RotateBasisTypeEnum } from '#app/schema/rotate'
 import { PanelFormDesignRotateEditValue } from '../../../forms/design/panel-form-design-rotate-edit-value'
 import { PanelPopoverDesignRotate } from '../../../popovers/design/panel-popover-design-rotate'
 
@@ -11,6 +12,16 @@ export const PanelDesignTypeRowValuesRotate = ({
 }: {
 	rotate: IRotate
 }) => {
+	const RotateBasisIcon = () => {
+		if (rotate.basis !== RotateBasisTypeEnum.DEFINED) return null
+
+		return (
+			<div className="m-2 mr-0 flex h-8 w-8 items-center justify-center">
+				<span className="text-body-xs leading-none">°</span>
+			</div>
+		)
+	}
+
 	return (
 		<SidebarPanelRowValuesContainer>
 			<PanelPopoverDesignRotate rotate={rotate} />
@@ -19,6 +30,7 @@ export const PanelDesignTypeRowValuesRotate = ({
 			) : (
 				<PanelFormDesignRotateEditValue rotate={rotate} />
 			)}
+			<RotateBasisIcon />
 		</SidebarPanelRowValuesContainer>
 	)
 }
