--- app/routes/sketch+/artboards+/$slug_+/components/panel/layer/design/panel-content-layer-design-stroke.tsx
@@ -1,143 +1,0 @@
-import {
-	Panel,
-	PanelHeader,
-	PanelRow,
-	PanelRowContainer,
-	PanelRowIconContainer,
-	PanelRowOrderContainer,
-	PanelRowValueContainer,
-	PanelTitle,
-} from '#app/components/shared'
-import { Input } from '#app/components/ui/input'
-import { type IDesignWithStroke } from '#app/models/design.server'
-import { type ILayer } from '#app/models/layer.server'
-import { DesignTypeEnum } from '#app/schema/design'
-import {
-	panelItemVariablesDesignType,
-	panelListVariablesDesignType,
-	selectedDesignsOnUpdate,
-} from '#app/utils/design'
-import { PanelFormDesignStrokeEditValue } from '../../../forms/design/panel-form-design-stroke-edit-value'
-import { PanelFormLayerDesignDelete } from '../../../forms/layer/design/panel-form-layer-design-delete'
-import { PanelFormLayerDesignNew } from '../../../forms/layer/design/panel-form-layer-design-new'
-import { PanelFormLayerDesignReorder } from '../../../forms/layer/design/panel-form-layer-design-reorder'
-import { PanelFormLayerDesignToggleVisibile } from '../../../forms/layer/design/panel-form-layer-design-toggle-visible'
-import { PanelPopoverDesignStroke } from '../../../popovers/design/panel-popover-design-stroke'
-
-export const PanelContentLayerDesignStroke = ({
-	layer,
-	designStrokes,
-}: {
-	layer: ILayer
-	designStrokes: IDesignWithStroke[]
-}) => {
-	const {
-		orderedDesignIds,
-		designCount,
-		visibleDesignIds,
-		firstVisibleDesignId,
-		selectedDesignId,
-	} = panelListVariablesDesignType({
-		designs: designStrokes,
-	})
-
-	return (
-		<Panel>
-			<PanelHeader>
-				<PanelTitle>Stroke</PanelTitle>
-				<div className="flex flex-shrink">
-					<PanelFormLayerDesignNew
-						layerId={layer.id}
-						type={DesignTypeEnum.STROKE}
-						visibleDesignsCount={visibleDesignIds.length}
-					/>
-				</div>
-			</PanelHeader>
-			{designStrokes.map((design, index) => {
-				const { id, visible, stroke } = design
-
-				const {
-					isSelectedDesign,
-					nextDesignId,
-					prevDesignId,
-					nextVisibleDesignId,
-				} = panelItemVariablesDesignType({
-					id,
-					selectedDesignId,
-					orderedDesignIds,
-					visibleDesignIds,
-				})
-
-				const {
-					selectDesignIdOnMoveUp,
-					selectDesignIdOnMoveDown,
-					selectDesignIdOnToggleVisible,
-					selectDesignIdOnDelete,
-				} = selectedDesignsOnUpdate({
-					id,
-					selectedDesignId,
-					isSelectedDesign,
-					visible,
-					prevDesignId,
-					nextDesignId,
-					nextVisibleDesignId,
-					firstVisibleDesignId,
-					orderedDesignIds,
-				})
-
-				return (
-					<PanelRow key={stroke.id}>
-						<PanelRowOrderContainer>
-							<PanelFormLayerDesignReorder
-								id={id}
-								layerId={layer.id}
-								panelCount={designCount}
-								panelIndex={index}
-								direction="up"
-								updateSelectedDesignId={selectDesignIdOnMoveUp}
-							/>
-							<PanelFormLayerDesignReorder
-								id={id}
-								layerId={layer.id}
-								panelCount={designCount}
-								panelIndex={index}
-								direction="down"
-								updateSelectedDesignId={selectDesignIdOnMoveDown}
-							/>
-						</PanelRowOrderContainer>
-						<PanelRowContainer>
-							<PanelRowValueContainer>
-								<PanelPopoverDesignStroke stroke={stroke} />
-								{/* this is a little buggy, but I can manage for now */}
-								{stroke.basis !== 'defined' ? (
-									<Input
-										type="text"
-										className={'flex h-8'}
-										disabled
-										defaultValue={stroke.basis}
-									/>
-								) : (
-									<PanelFormDesignStrokeEditValue stroke={stroke} />
-								)}
-							</PanelRowValueContainer>
-							<PanelRowIconContainer>
-								<PanelFormLayerDesignToggleVisibile
-									id={id}
-									layerId={layer.id}
-									visible={visible}
-									updateSelectedDesignId={selectDesignIdOnToggleVisible}
-								/>
-								<PanelFormLayerDesignDelete
-									id={id}
-									layerId={layer.id}
-									isSelectedDesign={false}
-									updateSelectedDesignId={selectDesignIdOnDelete}
-								/>
-							</PanelRowIconContainer>
-						</PanelRowContainer>
-					</PanelRow>
-				)
-			})}
-		</Panel>
-	)
-}
