--- app/routes/sketch+/artboards+/$slug_+/components/panel/layer/design/panel-content-layer-design-template.tsx
@@ -1,137 +1,0 @@
-import {
-	Panel,
-	PanelHeader,
-	PanelRow,
-	PanelRowContainer,
-	PanelRowIconContainer,
-	PanelRowOrderContainer,
-	PanelRowValueContainer,
-	PanelTitle,
-} from '#app/components/shared'
-import { Input } from '#app/components/ui/input'
-import { type IDesignWithTemplate } from '#app/models/design.server'
-import { type ILayer } from '#app/models/layer.server'
-import { DesignTypeEnum } from '#app/schema/design'
-import {
-	panelItemVariablesDesignType,
-	panelListVariablesDesignType,
-	selectedDesignsOnUpdate,
-} from '#app/utils/design'
-import { PanelFormLayerDesignDelete } from '../../../forms/layer/design/panel-form-layer-design-delete'
-import { PanelFormLayerDesignNew } from '../../../forms/layer/design/panel-form-layer-design-new'
-import { PanelFormLayerDesignReorder } from '../../../forms/layer/design/panel-form-layer-design-reorder'
-import { PanelFormLayerDesignToggleVisibile } from '../../../forms/layer/design/panel-form-layer-design-toggle-visible'
-import { PanelPopoverDesignTemplate } from '../../../popovers/design/panel-popover-design-template'
-
-export const PanelContentLayerDesignTemplate = ({
-	layer,
-	designTemplates,
-}: {
-	layer: ILayer
-	designTemplates: IDesignWithTemplate[]
-}) => {
-	const {
-		orderedDesignIds,
-		designCount,
-		visibleDesignIds,
-		firstVisibleDesignId,
-		selectedDesignId,
-	} = panelListVariablesDesignType({
-		designs: designTemplates,
-	})
-
-	return (
-		<Panel>
-			<PanelHeader>
-				<PanelTitle>Template</PanelTitle>
-				<div className="flex flex-shrink">
-					<PanelFormLayerDesignNew
-						layerId={layer.id}
-						type={DesignTypeEnum.TEMPLATE}
-						visibleDesignsCount={visibleDesignIds.length}
-					/>
-				</div>
-			</PanelHeader>
-			{designTemplates.map((design, index) => {
-				const { id, visible, template } = design
-
-				const {
-					isSelectedDesign,
-					nextDesignId,
-					prevDesignId,
-					nextVisibleDesignId,
-				} = panelItemVariablesDesignType({
-					id,
-					selectedDesignId,
-					orderedDesignIds,
-					visibleDesignIds,
-				})
-
-				const {
-					selectDesignIdOnMoveUp,
-					selectDesignIdOnMoveDown,
-					selectDesignIdOnToggleVisible,
-					selectDesignIdOnDelete,
-				} = selectedDesignsOnUpdate({
-					id,
-					selectedDesignId,
-					isSelectedDesign,
-					visible,
-					prevDesignId,
-					nextDesignId,
-					nextVisibleDesignId,
-					firstVisibleDesignId,
-					orderedDesignIds,
-				})
-
-				return (
-					<PanelRow key={template.id}>
-						<PanelRowOrderContainer>
-							<PanelFormLayerDesignReorder
-								id={id}
-								layerId={layer.id}
-								panelCount={designCount}
-								panelIndex={index}
-								direction="up"
-								updateSelectedDesignId={selectDesignIdOnMoveUp}
-							/>
-							<PanelFormLayerDesignReorder
-								id={id}
-								layerId={layer.id}
-								panelCount={designCount}
-								panelIndex={index}
-								direction="down"
-								updateSelectedDesignId={selectDesignIdOnMoveDown}
-							/>
-						</PanelRowOrderContainer>
-						<PanelRowContainer>
-							<PanelRowValueContainer>
-								<PanelPopoverDesignTemplate template={template} />
-								<Input
-									type="text"
-									className={'flex h-8'}
-									disabled
-									defaultValue={template.style}
-								/>
-							</PanelRowValueContainer>
-							<PanelRowIconContainer>
-								<PanelFormLayerDesignToggleVisibile
-									id={id}
-									layerId={layer.id}
-									visible={visible}
-									updateSelectedDesignId={selectDesignIdOnToggleVisible}
-								/>
-								<PanelFormLayerDesignDelete
-									id={id}
-									layerId={layer.id}
-									isSelectedDesign={false}
-									updateSelectedDesignId={selectDesignIdOnDelete}
-								/>
-							</PanelRowIconContainer>
-						</PanelRowContainer>
-					</PanelRow>
-				)
-			})}
-		</Panel>
-	)
-}
